[getVariantSettings] Rename file to match function name
