+++ src/lib/getVariantSettings.ts
